Chatbot for iP needs initial skeletal setup
Duke was the default setup for CS2103T's iP project, chatbot needs to be personalised

Rename Duke.java to Locky.java and update greeting and exit messages.

Locky is short and memorable and tied to the Gen-Z phrase "lock in" making it effective for a study to-do chatbot.

All traces of "Duke" are removed from the codebase to avoid confusion