The chatbot currently has no way to retain previous commands and list them out.
The chatbot should be able to store commands in a list and return them as prompted, in order to meet Duke Level-2 requirements.

Implement an ArrayList to store user-entered text, update the loop to add non-list commands to the list, and handle list command to print all stored tasks.

ArrayList has no fixed limit on the number of items in a list, hence can expand dynamically. Additional conditional should be implemented to add list feature.

Each added task is acknowledged with a confirmation message, and tasks are displayed in a numbered format when "list" is entered.