Add the ability to mark tasks as done or undone to Locky.
Mark tasks as done to meet Level-3 specs in CS2103T iP

Introduce a class Task with properties description and isDone, and use mark and unmark markers to mark a task as done.

Use a Task class to employ encapsulation to ensure code remains modular and external code does not interfere with internal.

Display tasks with internal toString() method with [X] and [] markers.